Disables MinVer package reference
Disables MinVer package reference.
When building the solution a pre-release version is always built. By commenting this package reference that logic is not applied there after. Added custom <VersionPrefix> and <VersionSuffix> tags to get a custom version of library.

Note :
Check "MinVer" for further details, did not go in deep yet on how to use it.